requirements.txt
numpy==1.23.5
opencv-python
opencv-contrib-python
pyyaml
onnx
torchvision 
onnxruntime
matplotlib
scipy
Cython
cmake
dlib
PyQt5
qt-thread-updater
pygame
PyOpenGL
PyOpenGL_accelerate